parser-service/requirements.txt
flask
psycopg2-binary
python-docx
PyMuPDF
openai
requests
tqdm
python-dotenv
spacy
ftfy